+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Karafka framework changelog
 
 ## 2.1.8 (Unreleased)
-- [Improvement] Mitigate a librdkafka `cooperative-sticky`.
+- [Improvement] Introduce `Karafka::BaseConsumer#used?` method to indicate, that at least one invocation of `#consume` took or will take place. This can be used as a replacement to the non-direct `messages.count` check for shutdown and revocation to ensure, that the consumption took place or is taking place (in case of running LRJ).
+- [Improvement] Make `messages#to_a` return copy of the underlying array to prevent scenarios, where the mutation impacts offset management.
+- [Improvement] Mitigate a librdkafka `cooperative-sticky` rebalance crash issue.
 - [Improvement] Provide ability to overwrite `consumer_persistence` per subscribed topic. This is mostly useful for plugins and extensions developers.
+- [Fix] Fix a case where the performance tracker would crash in case of mutation of messages to an empty state.
 
 ## 2.1.7 (2023-07-22)
 - [Improvement] Always query for watermarks in the Iterator to improve the initial response time.